Delete Dockerfile
To some how have a place where we can store our work and at anytime be able to pull the up when ever we need to revert back to something we might have forgotten . everything thing to me is personal .here. like a GitHub cloud  the belong only to us and not Google.